الحفظ

+++ InventoryWebApp/Patterns/Facade/InventoryFacade.cs
@@ -228,7 +228,7 @@ namespace InventoryWebApp.Patterns
                 if (comp.Price <= 0)
                     throw new Exception($"❌ سعر المكوّن {comp.ProductName} غير صالح");
 
-                int qtyPerUnit = componentQuantities[comp.Price];
+                decimal qtyPerUnit = componentQuantities[comp.Price];
 
                 // سعر المكوّن للوحدة
                 totalPrice += comp.Price * qtyPerUnit;
